% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Acupuncture}
\alias{Acupuncture}
\title{Acupuncture}
\format{A dataset with 1162 observations on the following 2 variables.
\describe{
  \item{Acupunture}{character}
  \item{Improvement}{character}
}}
\description{
A data set
}
\keyword{datasets}

