% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AgeBMI}
\alias{AgeBMI}
\title{AgeBMI}
\format{A dataset with 39 observations on the following 2 variables.
\describe{
  \item{Age}{integer}
  \item{bmi}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

