% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AlcoholSmoke}
\alias{AlcoholSmoke}
\title{AlcoholSmoke}
\format{A dataset with 863 observations on the following 2 variables.
\describe{
  \item{alcohol_drinks}{integer}
  \item{smoked}{integer}
}}
\description{
A data set
}
\keyword{datasets}

