% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Auction}
\alias{Auction}
\title{Auction}
\format{A dataset with 88 observations on the following 2 variables.
\describe{
  \item{dutch}{numeric}
  \item{FP}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

