% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Balsawood}
\alias{Balsawood}
\title{Balsawood}
\format{A dataset with 44 observations on the following 2 variables.
\describe{
  \item{treatment}{character}
  \item{elasticity}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

