% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BikeTimes}
\alias{BikeTimes}
\title{BikeTimes}
\format{A dataset with 56 observations on the following 2 variables.
\describe{
  \item{Frame}{character}
  \item{Time}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

