% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Blood}
\alias{Blood}
\title{Blood}
\format{A dataset with 2698 observations on the following 2 variables.
\describe{
  \item{year}{integer}
  \item{response}{character}
}}
\description{
A data set
}
\keyword{datasets}

