% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BMIGlucose}
\alias{BMIGlucose}
\title{BMIGlucose}
\format{A dataset with 38 observations on the following 2 variables.
\describe{
  \item{bmi}{numeric}
  \item{Glucose}{integer}
}}
\description{
A data set
}
\keyword{datasets}

