% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BowlsMMs}
\alias{BowlsMMs}
\title{BowlsMMs}
\format{A dataset with 17 observations on the following 2 variables.
\describe{
  \item{Small}{integer}
  \item{Large}{integer}
}}
\description{
A data set
}
\keyword{datasets}

