% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Brain}
\alias{Brain}
\title{Brain}
\format{A dataset with 107 observations on the following 2 variables.
\describe{
  \item{treatment}{character}
  \item{brain_change}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

