% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BreastFeedIntell}
\alias{BreastFeedIntell}
\title{BreastFeedIntell}
\format{A dataset with 322 observations on the following 2 variables.
\describe{
  \item{Feeding}{character}
  \item{GCI}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

