% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CarCrashDummies}
\alias{CarCrashDummies}
\title{CarCrashDummies}
\format{A dataset with 338 observations on the following 2 variables.
\describe{
  \item{doors}{character}
  \item{headinjury}{integer}
}}
\description{
A data set
}
\keyword{datasets}

