% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CarCrashDummiesLog}
\alias{CarCrashDummiesLog}
\title{CarCrashDummiesLog}
\format{A dataset with 338 observations on the following 2 variables.
\describe{
  \item{doors}{character}
  \item{logheadinjury}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

