% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CatJumping}
\alias{CatJumping}
\title{CatJumping}
\format{A dataset with 18 observations on the following 7 variables.
\describe{
  \item{id}{character}
  \item{sex}{character}
  \item{body_mass}{integer}
  \item{hind_limb_length}{numeric}
  \item{muscle_mass}{numeric}
  \item{body_fat}{integer}
  \item{takeoff_velocity}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

