% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Ch6Inv}
\alias{Ch6Inv}
\title{Ch6Inv}
\format{A dataset with 51 observations on the following 2 variables.
\describe{
  \item{Words}{character}
  \item{Memorized}{integer}
}}
\description{
A data set
}
\keyword{datasets}

