% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Ch9Inv}
\alias{Ch9Inv}
\title{Ch9Inv}
\format{A dataset with 72 observations on the following 2 variables.
\describe{
  \item{group}{character}
  \item{score}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

