% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ChildrenandLifespan}
\alias{ChildrenandLifespan}
\title{ChildrenandLifespan}
\format{A dataset with 90 observations on the following 2 variables.
\describe{
  \item{children}{character}
  \item{age}{integer}
}}
\description{
A data set
}
\keyword{datasets}

