% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cholesterol_day2_14}
\alias{Cholesterol_day2_14}
\title{Cholesterol_day2_14}
\format{A dataset with 19 observations on the following 2 variables.
\describe{
  \item{day2}{integer}
  \item{day14}{integer}
}}
\description{
A data set
}
\keyword{datasets}

