% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CloseFriends}
\alias{CloseFriends}
\title{CloseFriends}
\format{A dataset with 1467 observations on the following 2 variables.
\describe{
  \item{sex}{character}
  \item{friends}{integer}
}}
\description{
A data set
}
\keyword{datasets}

