% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CoffeeHeight}
\alias{CoffeeHeight}
\title{CoffeeHeight}
\format{A dataset with 51 observations on the following 2 variables.
\describe{
  \item{coffee}{character}
  \item{height}{integer}
}}
\description{
A data set
}
\keyword{datasets}

