% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Diets1}
\alias{Diets1}
\title{Diets1}
\format{A dataset with 232 observations on the following 2 variables.
\describe{
  \item{Diet}{character}
  \item{BMI}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

