% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Diets2}
\alias{Diets2}
\title{Diets2}
\format{A dataset with 93 observations on the following 2 variables.
\describe{
  \item{Diet}{character}
  \item{WeightLoss}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

