% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Donation}
\alias{Donation}
\title{Donation}
\format{A dataset with 44 observations on the following 2 variables.
\describe{
  \item{state}{character}
  \item{donation}{integer}
}}
\description{
A data set
}
\keyword{datasets}

