% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DraftLottery}
\alias{DraftLottery}
\title{DraftLottery}
\format{A dataset with 366 observations on the following 2 variables.
\describe{
  \item{sequential_date}{integer}
  \item{draft_number}{integer}
}}
\description{
A data set
}
\keyword{datasets}

