% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EarningsByDegree}
\alias{EarningsByDegree}
\title{EarningsByDegree}
\format{A dataset with 250 observations on the following 2 variables.
\describe{
  \item{Degree}{character}
  \item{Earnings..K.}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

