% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EcoliFlow}
\alias{EcoliFlow}
\title{EcoliFlow}
\format{A dataset with 23 observations on the following 1 variables.
\describe{
  \item{flow_ratio}{integer}
}}
\description{
A data set
}
\keyword{datasets}

