% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EcoliSand}
\alias{EcoliSand}
\title{EcoliSand}
\format{A dataset with 19 observations on the following 2 variables.
\describe{
  \item{sand_level}{character}
  \item{day_1}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

