% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ExamTimesScores}
\alias{ExamTimesScores}
\title{ExamTimesScores}
\format{A dataset with 30 observations on the following 2 variables.
\describe{
  \item{time}{integer}
  \item{score}{integer}
}}
\description{
A data set
}
\keyword{datasets}

