% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ExerciseMemorize}
\alias{ExerciseMemorize}
\title{ExerciseMemorize}
\format{A dataset with 31 observations on the following 2 variables.
\describe{
  \item{exercising}{integer}
  \item{notExercising}{integer}
}}
\description{
A data set
}
\keyword{datasets}

