% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ExerciseMood}
\alias{ExerciseMood}
\title{ExerciseMood}
\format{A dataset with 32 observations on the following 2 variables.
\describe{
  \item{exercise_intensity}{numeric}
  \item{change_mood}{integer}
}}
\description{
A data set
}
\keyword{datasets}

