% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Facebook}
\alias{Facebook}
\title{Facebook}
\format{A dataset with 40 observations on the following 2 variables.
\describe{
  \item{friends}{numeric}
  \item{density}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

