% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FirstBase}
\alias{FirstBase}
\title{FirstBase}
\format{A dataset with 22 observations on the following 2 variables.
\describe{
  \item{narrow}{numeric}
  \item{wide}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

