% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FlexibilityPre}
\alias{FlexibilityPre}
\title{FlexibilityPre}
\format{A dataset with 179 observations on the following 2 variables.
\describe{
  \item{sex}{character}
  \item{pre_flexibility}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

