% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FlexibilityPrePost}
\alias{FlexibilityPrePost}
\title{FlexibilityPrePost}
\format{A dataset with 179 observations on the following 2 variables.
\describe{
  \item{pre}{numeric}
  \item{post}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

