% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FlightTimes}
\alias{FlightTimes}
\title{FlightTimes}
\format{A dataset with 32 observations on the following 2 variables.
\describe{
  \item{WingLength}{integer}
  \item{FlightTime}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

