% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GettysburgAddress}
\alias{GettysburgAddress}
\title{GettysburgAddress}
\format{A dataset with 268 observations on the following 1 variables.
\describe{
  \item{word}{character}
}}
\description{
A data set
}
\keyword{datasets}

