% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Gilbert}
\alias{Gilbert}
\title{Gilbert}
\format{A dataset with 1641 observations on the following 2 variables.
\describe{
  \item{GilbertWorked.}{character}
  \item{Patient}{character}
}}
\description{
A data set
}
\keyword{datasets}

