% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GoodandBad}
\alias{GoodandBad}
\title{GoodandBad}
\format{A dataset with 30 observations on the following 2 variables.
\describe{
  \item{Wording}{character}
  \item{Perception}{character}
}}
\description{
A data set
}
\keyword{datasets}

