% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GPA}
\alias{GPA}
\title{GPA}
\format{A dataset with 42 observations on the following 2 variables.
\describe{
  \item{hours}{numeric}
  \item{gpa}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

