% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GrishamSentences}
\alias{GrishamSentences}
\title{GrishamSentences}
\format{A dataset with 84 observations on the following 2 variables.
\describe{
  \item{Book}{character}
  \item{Words}{integer}
}}
\description{
A data set
}
\keyword{datasets}

