% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HaircutCosts}
\alias{HaircutCosts}
\title{HaircutCosts}
\format{A dataset with 50 observations on the following 1 variables.
\describe{
  \item{cost}{integer}
}}
\description{
A data set
}
\keyword{datasets}

