% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Haircuts}
\alias{Haircuts}
\title{Haircuts}
\format{A dataset with 50 observations on the following 2 variables.
\describe{
  \item{Sex}{character}
  \item{Cost}{integer}
}}
\description{
A data set
}
\keyword{datasets}

