% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HaloMedals}
\alias{HaloMedals}
\title{HaloMedals}
\format{A dataset with 92 observations on the following 2 variables.
\describe{
  \item{level}{integer}
  \item{medals}{integer}
}}
\description{
A data set
}
\keyword{datasets}

