% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HandSpan}
\alias{HandSpan}
\title{HandSpan}
\format{A dataset with 45 observations on the following 2 variables.
\describe{
  \item{TootsieRolls}{integer}
  \item{handSpan}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

