% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Handwidth}
\alias{Handwidth}
\title{Handwidth}
\format{A dataset with 46 observations on the following 2 variables.
\describe{
  \item{hand_width}{numeric}
  \item{perceived_weight}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

