% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HealthcareAge}
\alias{HealthcareAge}
\title{HealthcareAge}
\format{A dataset with 2 observations on the following 4 variables.
\describe{
  \item{X18.29}{integer}
  \item{X30.49}{integer}
  \item{X50.64}{integer}
  \item{X65.}{integer}
}}
\description{
A data set
}
\keyword{datasets}

