% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HeartDiseaseBaldness}
\alias{HeartDiseaseBaldness}
\title{HeartDiseaseBaldness}
\format{A dataset with 1435 observations on the following 2 variables.
\describe{
  \item{Heart_Disease}{character}
  \item{Baldness}{character}
}}
\description{
A data set
}
\keyword{datasets}

