% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HeartDiseaseBaldness2}
\alias{HeartDiseaseBaldness2}
\title{HeartDiseaseBaldness2}
\format{A dataset with 1435 observations on the following 2 variables.
\describe{
  \item{heart_disease}{character}
  \item{baldness}{character}
}}
\description{
A data set
}
\keyword{datasets}

