% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HeartRates}
\alias{HeartRates}
\title{HeartRates}
\format{A dataset with 130 observations on the following 2 variables.
\describe{
  \item{Gender}{character}
  \item{HeartRate}{integer}
}}
\description{
A data set
}
\keyword{datasets}

