% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HeightBMI}
\alias{HeightBMI}
\title{HeightBMI}
\format{A dataset with 94 observations on the following 2 variables.
\describe{
  \item{BMI}{numeric}
  \item{height}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

