% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HeightHaircut}
\alias{HeightHaircut}
\title{HeightHaircut}
\format{A dataset with 94 observations on the following 2 variables.
\describe{
  \item{height}{numeric}
  \item{haircut}{integer}
}}
\description{
A data set
}
\keyword{datasets}

