% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HeightWeight}
\alias{HeightWeight}
\title{HeightWeight}
\format{A dataset with 28 observations on the following 2 variables.
\describe{
  \item{height}{integer}
  \item{weight}{integer}
}}
\description{
A data set
}
\keyword{datasets}

