% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HikeDistances}
\alias{HikeDistances}
\title{HikeDistances}
\format{A dataset with 72 observations on the following 2 variables.
\describe{
  \item{Distance}{numeric}
  \item{Time}{integer}
}}
\description{
A data set
}
\keyword{datasets}

