% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HikeElevations}
\alias{HikeElevations}
\title{HikeElevations}
\format{A dataset with 72 observations on the following 2 variables.
\describe{
  \item{ElevationGain}{integer}
  \item{Time}{integer}
}}
\description{
A data set
}
\keyword{datasets}

