% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Hockey}
\alias{Hockey}
\title{Hockey}
\format{A dataset with 44 observations on the following 1 variables.
\describe{
  \item{Goals}{integer}
}}
\description{
A data set
}
\keyword{datasets}

