% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HondaAgePrice}
\alias{HondaAgePrice}
\title{HondaAgePrice}
\format{A dataset with 90 observations on the following 2 variables.
\describe{
  \item{age}{integer}
  \item{price}{integer}
}}
\description{
A data set
}
\keyword{datasets}

