% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HousePrices}
\alias{HousePrices}
\title{HousePrices}
\format{A dataset with 20 observations on the following 2 variables.
\describe{
  \item{SqFt}{integer}
  \item{Price}{integer}
}}
\description{
A data set
}
\keyword{datasets}

