% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{legos}
\alias{legos}
\title{legos}
\format{A dataset with 157 observations on the following 2 variables.
\describe{
  \item{pieces}{integer}
  \item{price}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

