% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MajorPuzzle}
\alias{MajorPuzzle}
\title{MajorPuzzle}
\format{A dataset with 40 observations on the following 2 variables.
\describe{
  \item{major}{character}
  \item{time}{integer}
}}
\description{
A data set
}
\keyword{datasets}

