% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MaleTemp}
\alias{MaleTemp}
\title{MaleTemp}
\format{A dataset with 65 observations on the following 1 variables.
\describe{
  \item{maleTemp}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

