% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabulated.R
\docType{data}
\name{Marriage Views}
\alias{Marriage Views}
\alias{MarriageViews}
\title{Marriage Views}
\format{\preformatted{ num [1:2, 1:4] 236 300 313 416 401 745 68 143
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:2] "Yes" "No"
  ..$ : chr [1:4] "Millenial" "GenX" "Boomers" "Over 65"
}}
\usage{
MarriageViews
}
\description{
Marriage Views
}
\keyword{datasets}

