% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MarriageAges}
\alias{MarriageAges}
\title{MarriageAges}
\format{A dataset with 100 observations on the following 1 variables.
\describe{
  \item{diff}{integer}
}}
\description{
A data set
}
\keyword{datasets}

