% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MarriageAges1}
\alias{MarriageAges1}
\title{MarriageAges1}
\format{A dataset with 24 observations on the following 2 variables.
\describe{
  \item{husband}{integer}
  \item{wife}{integer}
}}
\description{
A data set
}
\keyword{datasets}

