% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Mercury}
\alias{Mercury}
\title{Mercury}
\format{A dataset with 912 observations on the following 2 variables.
\describe{
  \item{fish}{character}
  \item{mercuryLevel}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

