% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NightLight1}
\alias{NightLight1}
\title{NightLight1}
\format{A dataset with 2 observations on the following 3 variables.
\describe{
  \item{Darkness}{integer}
  \item{NightLight}{integer}
  \item{RoomLight}{integer}
}}
\description{
A data set
}
\keyword{datasets}

