% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{OldFaithful}
\alias{OldFaithful}
\title{OldFaithful}
\format{A dataset with 203 observations on the following 2 variables.
\describe{
  \item{year}{character}
  \item{time}{integer}
}}
\description{
A data set
}
\keyword{datasets}

