% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{OldFaithful1}
\alias{OldFaithful1}
\title{OldFaithful1}
\format{A dataset with 222 observations on the following 1 variables.
\describe{
  \item{Time}{integer}
}}
\description{
A data set
}
\keyword{datasets}

