% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{OldFaithful2}
\alias{OldFaithful2}
\title{OldFaithful2}
\format{A dataset with 222 observations on the following 2 variables.
\describe{
  \item{EruptionType}{character}
  \item{TimeBetween}{integer}
}}
\description{
A data set
}
\keyword{datasets}

