% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{OrganDonor}
\alias{OrganDonor}
\title{OrganDonor}
\format{A dataset with 161 observations on the following 2 variables.
\describe{
  \item{Default}{character}
  \item{Choice}{character}
}}
\description{
A data set
}
\keyword{datasets}

