% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PlateSize}
\alias{PlateSize}
\title{PlateSize}
\format{A dataset with 20 observations on the following 2 variables.
\describe{
  \item{year}{integer}
  \item{size}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

