% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Pop}
\alias{Pop}
\title{Pop}
\format{A dataset with 18000 observations on the following 3 variables.
\describe{
  \item{sleep1}{numeric}
  \item{sleep2}{numeric}
  \item{sleep3}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

