% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ReactionTimes}
\alias{ReactionTimes}
\title{ReactionTimes}
\format{A dataset with 20 observations on the following 2 variables.
\describe{
  \item{Dominant}{numeric}
  \item{Non.dominant}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

