% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Recall}
\alias{Recall}
\title{Recall}
\format{A dataset with 57 observations on the following 2 variables.
\describe{
  \item{Condition}{character}
  \item{Recall}{integer}
}}
\description{
A data set
}
\keyword{datasets}

