% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RestaurantMusic}
\alias{RestaurantMusic}
\title{RestaurantMusic}
\format{A dataset with 393 observations on the following 2 variables.
\describe{
  \item{musicType}{character}
  \item{amount}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

