% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RollerCoasters}
\alias{RollerCoasters}
\title{RollerCoasters}
\format{A dataset with 129 observations on the following 2 variables.
\describe{
  \item{speed}{integer}
  \item{height}{integer}
}}
\description{
A data set
}
\keyword{datasets}

