% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ScrabbleNames}
\alias{ScrabbleNames}
\title{ScrabbleNames}
\format{A dataset with 16 observations on the following 2 variables.
\describe{
  \item{Letters}{integer}
  \item{ScrabbleScore}{integer}
}}
\description{
A data set
}
\keyword{datasets}

