% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ScrabbleRatio}
\alias{ScrabbleRatio}
\title{ScrabbleRatio}
\format{A dataset with 16 observations on the following 2 variables.
\describe{
  \item{scrabbleScore}{integer}
  \item{ratio}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

