% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SingerHeights}
\alias{SingerHeights}
\title{SingerHeights}
\format{A dataset with 130 observations on the following 2 variables.
\describe{
  \item{part}{character}
  \item{height}{integer}
}}
\description{
A data set
}
\keyword{datasets}

