% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Sleep}
\alias{Sleep}
\title{Sleep}
\format{A dataset with 21 observations on the following 2 variables.
\describe{
  \item{sleep}{character}
  \item{time}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

