% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SleepGPA}
\alias{SleepGPA}
\title{SleepGPA}
\format{A dataset with 805 observations on the following 2 variables.
\describe{
  \item{Sleep}{numeric}
  \item{GPA}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

