% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SleepHeight}
\alias{SleepHeight}
\title{SleepHeight}
\format{A dataset with 95 observations on the following 2 variables.
\describe{
  \item{sleep}{numeric}
  \item{height}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

