% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SleepTimes}
\alias{SleepTimes}
\title{SleepTimes}
\format{A dataset with 22 observations on the following 1 variables.
\describe{
  \item{sleepHrs}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

