% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SmartphoneRace}
\alias{SmartphoneRace}
\title{SmartphoneRace}
\format{A dataset with 2 observations on the following 3 variables.
\describe{
  \item{White}{integer}
  \item{Black}{integer}
  \item{Hispanic}{integer}
}}
\description{
A data set
}
\keyword{datasets}

