% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Smoking}
\alias{Smoking}
\title{Smoking}
\format{A dataset with 4167 observations on the following 2 variables.
\describe{
  \item{Parents}{character}
  \item{Child}{character}
}}
\description{
A data set
}
\keyword{datasets}

