% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SnakeAge}
\alias{SnakeAge}
\title{SnakeAge}
\format{A dataset with 21 observations on the following 1 variables.
\describe{
  \item{age}{integer}
}}
\description{
A data set
}
\keyword{datasets}

