% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SocialMediaandSelfEsteem}
\alias{SocialMediaandSelfEsteem}
\title{SocialMediaandSelfEsteem}
\format{A dataset with 81 observations on the following 2 variables.
\describe{
  \item{media}{character}
  \item{score}{integer}
}}
\description{
A data set
}
\keyword{datasets}

