% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SocialNormsSex}
\alias{SocialNormsSex}
\title{SocialNormsSex}
\format{A dataset with 181 observations on the following 2 variables.
\describe{
  \item{experimenter}{character}
  \item{collision}{character}
}}
\description{
A data set
}
\keyword{datasets}

