% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Stop}
\alias{Stop}
\title{Stop}
\format{A dataset with 316 observations on the following 2 variables.
\describe{
  \item{position}{character}
  \item{stop}{character}
}}
\description{
A data set
}
\keyword{datasets}

