% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{StroopAgeTime}
\alias{StroopAgeTime}
\title{StroopAgeTime}
\format{A dataset with 50 observations on the following 2 variables.
\describe{
  \item{age}{integer}
  \item{time}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

