% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Study1}
\alias{Study1}
\title{Study1}
\format{A dataset with 27 observations on the following 2 variables.
\describe{
  \item{EV}{character}
  \item{RV}{integer}
}}
\description{
A data set
}
\keyword{datasets}

