% Generated by roxygen2 (4.0.2): do not edit by hand
\name{swap}
\alias{swap}
\title{swap values among columns of a data frame}
\usage{
swap(data, which)
}
\arguments{
\item{data}{a data frame}

\item{which}{a vector integer or character specifying columns in \code{data}}
}
\description{
swap values among columns of a data frame
}
\examples{
swap(KidsFeet, c("length","width"))
}

