% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TempHeart}
\alias{TempHeart}
\title{TempHeart}
\format{A dataset with 130 observations on the following 2 variables.
\describe{
  \item{BodyTemp}{numeric}
  \item{HeartRate}{integer}
}}
\description{
A data set
}
\keyword{datasets}

