% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TestTimes}
\alias{TestTimes}
\title{TestTimes}
\format{A dataset with 33 observations on the following 2 variables.
\describe{
  \item{time}{numeric}
  \item{score}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

