% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TextbookPrices}
\alias{TextbookPrices}
\title{TextbookPrices}
\format{A dataset with 30 observations on the following 2 variables.
\describe{
  \item{Price}{numeric}
  \item{Pages}{integer}
}}
\description{
A data set
}
\keyword{datasets}

