% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TimeEstimate}
\alias{TimeEstimate}
\title{TimeEstimate}
\format{A dataset with 48 observations on the following 1 variables.
\describe{
  \item{Estimate}{integer}
}}
\description{
A data set
}
\keyword{datasets}

