% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Tuna}
\alias{Tuna}
\title{Tuna}
\format{A dataset with 274 observations on the following 2 variables.
\describe{
  \item{tuna}{character}
  \item{mercury}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

