% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{UsedCars}
\alias{UsedCars}
\title{UsedCars}
\format{A dataset with 102 observations on the following 1 variables.
\describe{
  \item{price}{integer}
}}
\description{
A data set
}
\keyword{datasets}

