% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{VideoStress}
\alias{VideoStress}
\title{VideoStress}
\format{A dataset with 45 observations on the following 2 variables.
\describe{
  \item{type_of_video}{character}
  \item{Stress}{integer}
}}
\description{
A data set
}
\keyword{datasets}

