% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WeightHaircut}
\alias{WeightHaircut}
\title{WeightHaircut}
\format{A dataset with 93 observations on the following 2 variables.
\describe{
  \item{weight}{integer}
  \item{haircut}{integer}
}}
\description{
A data set
}
\keyword{datasets}

