% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Yawning}
\alias{Yawning}
\title{Yawning}
\format{A dataset with 50 observations on the following 2 variables.
\describe{
  \item{yawnSeed}{character}
  \item{response}{character}
}}
\description{
A data set
}
\keyword{datasets}

