% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{YawnTime}
\alias{YawnTime}
\title{YawnTime}
\format{A dataset with 2 observations on the following 4 variables.
\describe{
  \item{X8.11am}{integer}
  \item{X11am.2pm}{integer}
  \item{X2.5pm}{integer}
  \item{X5.8pm}{integer}
}}
\description{
A data set
}
\keyword{datasets}

