% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BodyFat}
\alias{BodyFat}
\title{BodyFat}
\format{A dataset with 177 observations on the following 2 variables.
\describe{
  \item{pre_body_fat}{numeric}
  \item{post_body_fat}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

