% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CerealSugar}
\alias{CerealSugar}
\title{CerealSugar}
\format{A dataset with 80 observations on the following 2 variables.
\describe{
  \item{shelves}{character}
  \item{sugar}{integer}
}}
\description{
A data set
}
\keyword{datasets}

