% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CollegeMidwest}
\alias{CollegeMidwest}
\title{CollegeMidwest}
\format{A dataset with 2919 observations on the following 2 variables.
\describe{
  \item{onCampus}{character}
  \item{cumGPA}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

