% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DadandChildHeights}
\alias{DadandChildHeights}
\title{DadandChildHeights}
\format{A dataset with 25 observations on the following 2 variables.
\describe{
  \item{DadHeight}{integer}
  \item{Height}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

