% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EcoliTime}
\alias{EcoliTime}
\title{EcoliTime}
\format{A dataset with 14 observations on the following 2 variables.
\describe{
  \item{day_1}{numeric}
  \item{day_2}{integer}
}}
\description{
A data set
}
\keyword{datasets}

