% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GestationLifeExpectancy}
\alias{GestationLifeExpectancy}
\title{GestationLifeExpectancy}
\format{A dataset with 10 observations on the following 2 variables.
\describe{
  \item{Gest}{integer}
  \item{LifeExpct}{integer}
}}
\description{
A data set
}
\keyword{datasets}

