% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HeightandFingerLength}
\alias{HeightandFingerLength}
\title{HeightandFingerLength}
\format{A dataset with 34 observations on the following 2 variables.
\describe{
  \item{finger}{numeric}
  \item{height}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

