% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LiarLiar}
\alias{LiarLiar}
\title{LiarLiar}
\format{A dataset with 86 observations on the following 2 variables.
\describe{
  \item{relationship}{character}
  \item{number_correct}{integer}
}}
\description{
A data set
}
\keyword{datasets}

