% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Milwaukee}
\alias{Milwaukee}
\title{Milwaukee}
\format{A dataset with 69 observations on the following 2 variables.
\describe{
  \item{City}{character}
  \item{Estimate}{integer}
}}
\description{
A data set
}
\keyword{datasets}

