% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MomandChildHeights}
\alias{MomandChildHeights}
\title{MomandChildHeights}
\format{A dataset with 26 observations on the following 2 variables.
\describe{
  \item{momHeight}{integer}
  \item{height}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

