% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{OrganDonor}
\alias{OrganDonor}
\title{Organ Donation}
\format{A dataset with 161 observations on the following variables.
\itemize{
\item{\code{default}}
  {\code{opt-in}, \code{opt-out}, or \code{neutral}}
\item{\code{donor}}
  {\code{donor} or \code{not}}
  }}
\description{
How becoming an organ donor depends on how a person is asked.
}
\keyword{datasets}

