% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WaterWeight}
\alias{WaterWeight}
\title{WaterWeight}
\format{A dataset with 178 observations on the following 2 variables.
\describe{
  \item{pre_water_weight}{numeric}
  \item{post_water_weight}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

