% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Weight}
\alias{Weight}
\title{Weight}
\format{A dataset with 180 observations on the following 2 variables.
\describe{
  \item{pre_weight}{numeric}
  \item{post_weight}{numeric}
}}
\description{
A data set
}
\keyword{datasets}

